requirements.txt
python_telegram_bot==13.7
aiohttp==3.7.4.post0
schedule==1.1.0
gql==3.0.0a6
psycopg2_binary==2.9.1
python-dotenv==0.19.0
